Updated security scripts

+++ app/Http/Controllers/V1/AlbumController.php
@@ -2,20 +2,21 @@
 
 namespace App\Http\Controllers\V1;
 
-use App\Http\Controllers\Controller;
 use App\Models\Album;
+use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreAlbumRequest;
-use App\Http\Requests\UpdateAlbumRequest;
 use App\Http\Resources\V1\AlbumResource;
+use App\Http\Requests\UpdateAlbumRequest;
 
 class AlbumController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-        return AlbumResource::collection(Album::paginate(20));
+        return AlbumResource::collection(Album::where("user_id", $request->user()->id)->paginate(20));
     }
 
     /**
@@ -23,28 +24,32 @@ class AlbumController extends Controller
      */
     public function store(StoreAlbumRequest $request)
     {
-        $album = Album::create($request->all());
+        $data = $request->all();
+        $data["user_id"] = $request->user()->id;
+        $album = Album::create($data);
         return new AlbumResource($album);
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(Album $album)
+    public function show(Request $request, Album $album)
     {
-        if(is_null($album->name)){
-            return response()->json(["message"=>"This album does not exist."], 404);
-        }else{
-            return new AlbumResource($album);
+        if($request->user()->id != $album->user_id){
+            return abort(403, "Unauthorized");
         }
-        
+        return new AlbumResource($album);
+            
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateAlbumRequest $request, Album $album)
-    {
+    public function update(UpdateAlbumRequest $request, Album $album){
+    
+        if($request->user()->id != $album->user_id){
+            return abort(403, "Unauthorized");
+        }
         $album->update($request->all());
         return new AlbumResource($album);
     }
@@ -52,8 +57,11 @@ class AlbumController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Album $album)
+    public function destroy(Request $request, Album $album)
     {
+        if($request->user()->id != $album->user_id){
+            return abort(403, "Unauthorized");
+        }
         $album->delete();
         return response("", 204);
     }
