+++ app/Http/Controllers/V1/ImageManipulationController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\V1;
 
 use App\Models\Album;
 use Illuminate\Support\Str;
+use Illuminate\Http\Request;
 use App\Models\ImageManipulation;
 use Illuminate\Http\UploadedFile;
 use App\Http\Controllers\Controller;
@@ -17,12 +18,15 @@ class ImageManipulationController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-       return ImageManipulationResource::collection(ImageManipulation::paginate());
+       return ImageManipulationResource::collection(ImageManipulation::where("user_id", $request->user()->id)->paginate());
     }
 
-    public function byAlbum($album){
+    public function byAlbum(Request $request, $album){
+        if($request->user()->id != $album->user_id){
+            return abort(403, "Unauthorized");
+        }
         $where = [
             "album_id" => $album
         ];
@@ -40,10 +44,14 @@ class ImageManipulationController extends Controller
         $data = [
             "type" => ImageManipulation::TYPE_RESIZE,
             "data" => json_encode($all),
-            "user_id" => null
+            "user_id" => $request->user()->id
         ];
         if(isset($all["album_id"])){
-        $data["album_id"] = $all["album_id"];
+            $album = Album::find($all["album_id"]);
+            if($request->user()->id != $album->user_id){
+                return abort(403, "Unauthorized");
+            }
+            $data["album_id"] = $all["album_id"];
         }
 
         $dir = "images/" . Str::random() . "/";
@@ -80,16 +88,22 @@ class ImageManipulationController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(ImageManipulation $image)
+    public function show(Request $request, ImageManipulation $image)
     {
+        if($request->user()->id != $image->user_id){
+            return abort(403, "Unauthorized");
+        }
         return new ImageManipulationResource($image);
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(ImageManipulation $image)
+    public function destroy(Request $request, ImageManipulation $image)
     {
+        if($request->user()->id != $image->user_id){
+            return abort(403, "Unauthorized");
+        }
         $image->delete();
         return response("", 204);
     }
